add bulk message feature + initSessions

+++ src/whatsapp.ts
@@ -27,8 +27,8 @@ const msgRetryCounterCache = new NodeCache();
 // }, 10_000);
 
 export async function LogoutDevice(number: string, io: Socket) {
-  const session = sessions.get(number)
-  await session?.logout()
+  const session = sessions.get(number);
+  await session?.logout();
   const device = await prisma.numbers.findFirst({
     where: { body: number },
   });
@@ -39,11 +39,12 @@ export async function LogoutDevice(number: string, io: Socket) {
       data: { status: "Disconnect" },
     });
   }
-  fs.rmdirSync(`./${number}`, { recursive: true });
-  sessions.delete(number)  
-  connectToWhatsApp(number, io)
+  if (fs.existsSync(`./${number}`)) {
+    fs.rmdirSync(`./${number}`, { recursive: true });
+  }
+  sessions.delete(number);
+  connectToWhatsApp(number, io);
 }
-
 
 export async function connectToWhatsApp(number: string, io: Socket) {
   logger.info("SOCKET READY");
@@ -61,8 +62,6 @@ export async function connectToWhatsApp(number: string, io: Socket) {
     msgRetryCounterCache,
     generateHighQualityLinkPreview: true,
   });
-  
-  
 
   // store?.bind(sock.ev);
   sock.ev.process(
@@ -76,7 +75,20 @@ export async function connectToWhatsApp(number: string, io: Socket) {
         });
         const update = events["connection.update"];
         const { connection, lastDisconnect, qr } = update;
+
         if (qr?.length) {
+          logger.warn("QRCODE");
+          console.log(qr);
+          if (
+            (device?.status === "Connected" &&
+              update.connection === "connecting") ||
+            (device?.status === "Connected" && update.connection === "close")
+          ) {
+            await prisma.numbers.update({
+              where: { id: device.id },
+              data: { status: "Disconnect" },
+            });
+          }
           let qrcode = await toDataURL(qr);
           io.emit("qrcode", {
             token: number,
@@ -98,7 +110,7 @@ export async function connectToWhatsApp(number: string, io: Socket) {
             logger.error("PROFILE NOT FOUND");
           }
           if (result.jid.replace(/\D/g, "") != number.toString()) {
-            io.emit("number-mismatch")
+            io.emit("number-mismatch");
             await sock.logout();
           } else {
             io.emit("connection-open", {
@@ -148,7 +160,7 @@ export async function connectToWhatsApp(number: string, io: Socket) {
       // credentials updated -- save them
       if (events["creds.update"]) {
         await saveCreds();
-      }      
+      }
       // received a new message
       // if (events["messages.upsert"]) {
       //   const upsert = events["messages.upsert"];
@@ -168,3 +180,105 @@ export async function connectToWhatsApp(number: string, io: Socket) {
 
   sessions.set(number, sock);
 }
+
+export const initializeWhatsapp = async (number: string) => {
+  logger.info("SOCKET READY");
+  const { state, saveCreds } = await useMultiFileAuthState(`${number}`);
+  const { version, isLatest } = await fetchLatestBaileysVersion();
+  logger.info(`using WA v${version.join(".")}, isLatest: ${isLatest}`);
+  const sock = makeWASocket({
+    // can provide additional config here
+    version,
+    printQRInTerminal: true,
+    auth: {
+      creds: state.creds,
+      keys: makeCacheableSignalKeyStore(state.keys, logger),
+    },
+    msgRetryCounterCache,
+    generateHighQualityLinkPreview: true,
+  });
+
+  // store?.bind(sock.ev);
+  sock.ev.process(
+    // events is a map for event name => event data
+    async (events) => {
+      // something about the connection changed
+      // maybe it closed, or we received all offline message or connection opened
+      if (events["connection.update"]) {
+        const device = await prisma.numbers.findFirst({
+          where: { body: number },
+        });
+        const update = events["connection.update"];
+        const { connection, lastDisconnect, qr } = update;
+
+        if (qr?.length) {
+          logger.warn("QRCODE");
+          console.log(qr);
+          if (
+            (device?.status === "Connected" &&
+              update.connection === "connecting") ||
+            (device?.status === "Connected" && update.connection === "close")
+          ) {
+            await prisma.numbers.update({
+              where: { id: device.id },
+              data: { status: "Disconnect" },
+            });
+          }
+        }
+
+        if (connection === "open") {
+          await prisma.numbers.update({
+            where: { id: device?.id },
+            data: { status: "Connected" },
+          });
+          const [result] = await sock.onWhatsApp(sock.user?.id ?? "");
+          let ppUrl;
+          try {
+            ppUrl = await sock.profilePictureUrl(result.jid, "image");
+          } catch (error) {
+            logger.error("PROFILE NOT FOUND");
+          }
+        }
+        if (connection === "close") {
+          // reconnect if not logged out
+          if ((lastDisconnect?.error as Boom)?.output.statusCode === 515) {
+            initializeWhatsapp(number);
+          }
+          if (
+            (lastDisconnect?.error as Boom)?.output?.statusCode !==
+            DisconnectReason.loggedOut
+          ) {
+            if (device?.status === "Connected") {
+              await prisma.numbers.update({
+                where: { id: device.id },
+                data: { status: "Disconnect" },
+              });
+            }
+            // console.log(lastDisconnect?.error?.name)
+            // connectToWhatsApp(`${number}`, io);
+          } else {
+            fs.rmdirSync(`./${number}`, { recursive: true });
+            initializeWhatsapp(number);
+            const device = await prisma.numbers.findFirst({
+              where: { body: number },
+            });
+            await prisma.numbers.update({
+              where: { id: device?.id },
+              data: { status: "Disconnect" },
+            });
+            console.log("Connection closed. You are logged out.");
+          }
+        }
+
+        console.log("CONNECTION UPDATE", update);
+      }
+
+      // credentials updated -- save them
+      if (events["creds.update"]) {
+        await saveCreds();
+      }
+    }
+  );
+
+  sessions.set(number, sock);
+};
